# api/openapi.yaml
openapi: 3.1.0
info:
  title: Automotive Repair Workshop Management API
  description: >
    REST API cho hệ thống quản lý gara sửa chữa ô tô.
    Luồng chính: Tiếp nhận xe → Chẩn đoán → Báo giá → Duyệt → Sửa chữa (trừ kho phụ tùng)
    → Hoàn thành → Hóa đơn → Thanh toán.
  version: 1.0.0
servers:
  - url: http://localhost:3000/api/v1
    description: Local development
...
paths:
...
  /repair-jobs/{id}/assign:
    put:
      tags: [RepairJobs]
      summary: Phân công thợ máy (US07)
      operationId: assignMechanic
      security: [{ bearerAuth: [] }]
      parameters:
        - { $ref: '#/components/parameters/IdPath' }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [mechanicId]
              properties:
                mechanicId: { type: integer, format: int64 }
      responses:
        '200': { description: Đã phân công, phiếu chuyển sang diagnosing }
        '403': { $ref: '#/components/responses/Forbidden' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    IdPath:
      in: path
      name: id
      required: true
      schema: { type: integer, format: int64 }
  responses:
...
    Forbidden:
      description: Sai phân quyền (RBAC)
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
  schemas:
    Error:
      type: object
      properties:
        success: { type: boolean, const: false }
        message: { type: string }
...